Splits git utils into webview safe vs vscode only Moves paused operation status strings into utils
