+++ src/commands/quickCommand.steps.ts
@@ -43,8 +43,8 @@ import type { WorktreeQuickPickItem } from '../git/models/worktree.quickpick';
 import { createWorktreeQuickPickItem } from '../git/models/worktree.quickpick';
 import { getWorktreesByBranch } from '../git/models/worktree.utils';
 import { remoteUrlRegex } from '../git/parsers/remoteParser';
-import type { BranchSortOptions, TagSortOptions } from '../git/utils/sorting';
-import { sortBranches, sortContributors, sortTags, sortWorktrees } from '../git/utils/sorting';
+import type { BranchSortOptions, TagSortOptions } from '../git/utils/vscode/sorting';
+import { sortBranches, sortContributors, sortTags, sortWorktrees } from '../git/utils/vscode/sorting';
 import { getApplicablePromo } from '../plus/gk/account/promos';
 import { isSubscriptionPaidPlan, isSubscriptionPreviewTrialExpired } from '../plus/gk/account/subscription';
 import type { LaunchpadCommandArgs } from '../plus/launchpad/launchpad';
