+++ src/git/utils/vscode/sorting.ts
@@ -1,15 +1,15 @@
-import type { BranchSorting, ContributorSorting, RepositoriesSorting, TagSorting } from '../../config';
-import { sortCompare } from '../../system/string';
-import { configuration } from '../../system/vscode/configuration';
-import type { GitBranch } from '../models/branch';
-import type { GitContributor } from '../models/contributor';
-import { isContributor } from '../models/contributor';
-import type { ContributorQuickPickItem } from '../models/contributor.quickpick';
-import type { Repository } from '../models/repository';
-import type { GitTag } from '../models/tag';
-import type { GitWorktree } from '../models/worktree';
-import { isWorktree } from '../models/worktree';
-import type { WorktreeQuickPickItem } from '../models/worktree.quickpick';
+import type { BranchSorting, ContributorSorting, RepositoriesSorting, TagSorting } from '../../../config';
+import { sortCompare } from '../../../system/string';
+import { configuration } from '../../../system/vscode/configuration';
+import type { GitBranch } from '../../models/branch';
+import type { GitContributor } from '../../models/contributor';
+import { isContributor } from '../../models/contributor';
+import type { ContributorQuickPickItem } from '../../models/contributor.quickpick';
+import type { Repository } from '../../models/repository';
+import type { GitTag } from '../../models/tag';
+import type { GitWorktree } from '../../models/worktree';
+import { isWorktree } from '../../models/worktree';
+import type { WorktreeQuickPickItem } from '../../models/worktree.quickpick';
 
 export interface BranchSortOptions {
 	current?: boolean;
