+++ src/plus/integrations/providers/github/githubGitProvider.ts
@@ -88,9 +88,9 @@ import type { GitWorktree } from '../../../../git/models/worktree';
 import { getRemoteProviderMatcher, loadRemoteProviders } from '../../../../git/remotes/remoteProviders';
 import type { GitSearch, GitSearchResultData, GitSearchResults } from '../../../../git/search';
 import { getSearchQueryComparisonKey, parseSearchQuery } from '../../../../git/search';
-import { getRemoteIconUri } from '../../../../git/utils/icons';
-import type { BranchSortOptions, TagSortOptions } from '../../../../git/utils/sorting';
-import { sortBranches, sortTags } from '../../../../git/utils/sorting';
+import { getRemoteIconUri } from '../../../../git/utils/vscode/icons';
+import type { BranchSortOptions, TagSortOptions } from '../../../../git/utils/vscode/sorting';
+import { sortBranches, sortTags } from '../../../../git/utils/vscode/sorting';
 import { gate } from '../../../../system/decorators/gate';
 import { debug, log } from '../../../../system/decorators/log';
 import { filterMap, first, last, map, some, union } from '../../../../system/iterable';
